requirements-dev.txt
# Core libraries used by RewordApp
pyyaml

# Development & testing
pytest